requirements.txt
torch~=1.6.0
torchvision~=0.7.0
apex
numpy~=1.19.1
pillow~=7.2.0
setuptools~=49.6.0
matplotlib~=3.3.1
scipy~=1.5.4
dominate~=2.6.0
opencv-python~=4.4.0.46
pytz~=2020.5
tqdm~=4.48.2
requests~=2.24.0